refactor(domain): rename TrackedByWeight to TrackedByMeasure and implement universal SAT dictionary
- Renamed ProductType enum value to TrackedByMeasure to support weight, volume, length, and area.
- Added and applied idempotent data migration 20260518065026_RenameTrackedByWeightToMeasure.
- Enriched KdsItem record and 7 projection surfaces (Order Pull, Active Tables, Deliveries) with SatUnitCode.
- Implemented high-performance O(1) SatUnitSymbols dictionary for 26 UN/ECE Rec 20 continuous and discrete codes.
- Refactored ESC/POS ticket engine to use strict, zero-fallback unified rendering format.

+++ POS.Domain/Models/DeductSaleRequest.cs
@@ -15,5 +15,5 @@ public class DeductSaleRequest
 public class SaleItem
 {
     public int ProductId { get; set; }
-    public int Quantity { get; set; }
+    public decimal Quantity { get; set; }
 }
